backend/pyproject.toml
[project]
name = "dg-helper-backend"
version = "0.1.0"
description = "FastAPI backend for dg-helper"
requires-python = ">=3.12"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "python-jose[cryptography]",
    "passlib"
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "flake8",
    "mypy",
    "pytest",
    "pytest-asyncio",
    "httpx"
]
test = [
    "pytest",
    "pytest-asyncio",
    "httpx"
]

[build-system]
requires = ["uv"]
build-backend = "uv"
